Зайцев Артем. Задача 2. Вариант 8. Обобщенная передача от одного всем (scatter). (#370)
### The problem 
Finding the minimum element of a vector
### Algorithm
1. Divide the vector into sub-vectors and distribute them among the
threads.
2. Find the local minimums of each subvector.
3. Find the global minimum among local ones.
### Implementation
1. Signatures of the self-written `scatter(...)` and `MPI_Scatter(...)`
are identical.
2. If it's impossible to distribute the elements of the vector equally
between the processes, the unallocated values from the end of the vector
will be taken into account and processed in the root process.
3. `class TestMPITaskParallel<T, func>(...)` is a template: when
creating an instance, you must specify the type `T` of values in the
vector and the function `func` that distributes data to processes. The
tests consider the options `T ∈ {int, double}` and `func ∈ {scatter,
MPI_Scatter}`.